Version 0.8.4: Deleted: - Holy Cloud Entity - Holy Cloud Egg Block - Holy Cloud Spawn Egg - Holy Seeds
Changes:
- Holy Weapon Model
- Holy Weapon Item Texture
- Holy Weapon Craft
- Holy Weapon Name: Holy Weapon → Divine Thunderbolt

Explanation:
Holy Cloud:
This mod is for Roleplay and I think that a useless Mob it doesn’t fit there. Also, working with this mob is making me a lot of problems as animations and AI.

Holy Weapon:
I wanted to change the item to fit with the Divine Gatherer. As it is a cloud and the item throws lightnings, I changed the model to a lightning.